requirements.txt
# requirements.txt

aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bs4==0.0.1
cchardet==2.1.7
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dearpygui==1.10.1
docxcompose==1.4.0
docxtpl==0.16.7
frozenlist==1.4.0
geographiclib==2.0
geopy==2.4.0
greenlet==3.0.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
nltk==3.8.1
people-also-ask==1.1.0
Pillow==10.1.0
playwright==1.39.0
pyee==11.0.1
pyoxidizer==0.24.0
python-docx==1.1.0
readability==0.3.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
xdialog==1.1.1.1
yarl==1.9.2
nltk
